refactor(hush-cli): extract command from pkg_cli
Move the four clap arg-type enums (CliPkgType, PkgCommands,
TrustedPublisherCommands, OrgCommands) into pkg_cli/command.rs.
CliPkgType::{to_pkg_type,label} widened fn -> pub(super) for the
init handler's cross-module access. Re-exported from mod.rs to
preserve the public path pkg_cli::{CliPkgType, OrgCommands,
PkgCommands, TrustedPublisherCommands}.
